Do not start line with number, this turns into ol
Do not start the line with a number, as github pages turns this into an ordered list, such that the number 2019 now does not occur at all on http://pytroll.github.io/pytroll_packages_overview.html and instead an ordered list starts: "is not guaranteed after \ \ 1. If a package..."  Moving the word `after` to the next line should fix this (although I don't know how to test this).