Small updates for coding guidelines
In the coding guidelines, remove a reference to Python 2 support, which
is of historical interest only.  Replace a reference to stickler by a
more generic "automated tools" description.  Remove a reference that
people can use git format-patch to the bdfl, which I think nobody does
and I don't think the bdfls want, and I'm not sure if we want to use the
term bdfl anyway.